pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "osads"
version = "0.1.0"
description = "Open Source Air Defense System - Insect/Drone detection, tracking & targeting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "opencv-python>=4.8",
    "scipy>=1.11",
    "PyYAML>=6.0",
    "torch>=2.0",
    "torchaudio>=2.0",
    "torchvision>=0.15",
    "ultralytics>=8.0",
    "librosa>=0.10",
    "soundfile>=0.12",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.5",
    "ruff>=0.1",
]
dashboard = [
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "jinja2>=3.1",
]
hardware = [
    "pigpio>=1.78",
    "picamera2>=0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.ruff]
target-version = "py311"
line-length = 100
